requirements.txt
alembic==1.7.6
bcrypt==3.2.0
bidict==0.21.4
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.2
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
